chore: enable container security contexts modifications
---

Resolves issue: https://github.com/gomods/athens-charts/issues/49.

This commit updates the chart to enable us to override/set additional
security context configuration at the container level in each of the
containers in the athens-proxy deployment template.

It moves the existing `securityContext` configuration to the
`PodSecurityContext` at the pod spec level to ensure this isn't a
breaking change and ensures backwards compatibility. It adds two new
values in the chart `securityContext` (for the main athens container)
and `initContainerSecurityContext` (for the init container, if used).
This allows us to override/modify the existing configuration with
additional security context configuration options in each of the
containers, at the container `securityContext` level.

See `PodSecurityContext` API reference for fields that can be set at the
pod spec level (for the existing `image.runAsNonRoot` configuration):
https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.28/#podsecuritycontext-v1-core.

And see `SecurityContext` API reference for the fields that can be set
at the container spec level:
https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.28/#securitycontext-v1-core.