fix: preprovisioned service account unable to be included in deployment (#74)
#71 excludes serviceAccounts in the deployment template entirely if
`serviceAccount.enabled` is set to false, meaning it defaults to
"default" for the deployment. Previously, we could create a service
account that was named identical to `fullname`, i.e., either the release
name or an override via values, and have that SA be used for the pod.

This PR re-adds the ability to use service accounts in the deployment
*without* creating one, and also extends upon the pre-#71 behavior of
allowing you to set the name of the service account to use via
`serviceAccount.name`. The property is not set in the chart's default
values, and defaults to the `fullname` behavior of < v0.9.4.

Additionally, we also swapped out a bunch of "templates" for "includes",
as per
https://helm.sh/docs/howto/charts_tips_and_tricks/#using-the-include-function.

---------

Signed-off-by: Alexander Bartolomey <github@alexanderbartolomey.de>