Addition of configurable init containers to athens-proxy
Resolves issue #106

This adds the ability to add extra init containers to the athens proxy. I'm using a credentials-helper to provide authentication to the git repositories for athens. Using an init container can ensure the volumes are set up correctly.

Also adding clarification to the existing initContainer configuration 'initContainerSecurityContext' and 'intiContainerResources' to indicate they only apply to the 'sshGitServer' init container.